Moved FixedScansController to its own class
