+++ vimms/Controller/__init__.py
@@ -4,3 +4,4 @@ from vimms.Controller.topN import *
 from vimms.Controller.roi import *
 from vimms.Controller.tree import *
 from vimms.Controller.dia import *
+from vimms.Controller.misc import *
